atualização da view de artigo e usuário: realocação do metodo de check_role
